Delete starred messages only if the table exists

+++ src/main/kotlin/data/db/StarredMessageDao.kt
@@ -82,7 +82,9 @@ class StarredMessageDao {
 
     fun deleteAll() {
         transaction {
-            StarredMessages.deleteAll()
+            if (StarredMessages.exists()) {
+                StarredMessages.deleteAll()
+            }
         }
     }
